Smooth-based evolution is also affected by convergence. When we are close to the final solution the smoothing term is damped by a functional that goes to zero when the distance term is zero.
The gridaccell has also been modified to march rays not only in the direction of the ray itself, 
but also up to an offset in the opposite direction. This can be done by offsetting the starting
point along -r.d by some value. GridAccell::trace_ray( ray, off ).

Note that The vasewidget settings have been changed for this new debugging step!!



git-svn-id: svn://svn.code.sf.net/p/meshlab/code/trunk@4229 324bb2ba-784c-0410-a299-e66b5c357fcb