Major: removed any reference to the original code from Lior Shapira. Moved all implementation directly in the plugin body (filter_sdf.cpp).
Note: will work only on vertices, for faces is not yet implemented. It would make sense to have the plugin feed on a mesh document made of two elements: 1) a mesh 2) a sample cloud on the mesh In this way you could also use any sampling algorithm that meshlab has.

Next step: generation of medial "cloud" and back-correction

git-svn-id: svn://svn.code.sf.net/p/meshlab/code/trunk@4291 324bb2ba-784c-0410-a299-e66b5c357fcb