Very first commit od the OCME plugin under MeshLab.
what it does:
allows to put together an arbitrary number of meshes in a new editable data structure
that preserves all the original information .
The data structure is named OCM (Out of Core Mesh) and can be visually browsed in MeshLab 
with a simple impostor based visualization. 

OCME (OCM Editor) allows you to modify the data with a add/edit/commit paradigm:
you can ADD meshes to the data structure;
you can select a portion of the data that can be entirely loaded in memory as a MeshLab 
layer (EDIT), so that you can apply all the meshlab filter, and then put it back the result (COMMIT).

NOTE:
- there are NO external dependencies, although you may find "BerkeleyDB" in many places, it is an alternative
compilation path which is currently excluded

- The number of meshes you can put in a single OCM object is not that arbitrary. At the present the simple database
structure does not allow that, with 2G of RAM you should be fine with 1G triangles more or less.

- commit is not enabled yet, meaning that you can build the dataset by adding meshes, browse it, take pieces to edit but not commit  it back,
therefore it is just for watching/playing
- it works under linux ut has not been recompilaed under windows recently

- under edit_ocme/src/ocme there is the project ocme.pro which produces a command line builder to build a OCM data structure from
 a bunch of ply files or from an aln file. 




git-svn-id: svn://svn.code.sf.net/p/meshlab/code/trunk@4393 324bb2ba-784c-0410-a299-e66b5c357fcb